+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added unique object ids for SVG output object tags `<g>`
 - Added support for global CSS styling using the `--global-style` option in the CLI
 - Added optional attributes width and height for each drawn object
+- Added `--theme` option to the CLI to apply stylized themes
 
 ### 🐛 Bug fixes
 
@@ -41,6 +42,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Increased timeout for CLI tests from 2 seconds to 30 seconds
 - Added appropriate icons to the buttons on the demo and webstepper pages
 - Updated demo page tests in accordance with this change
+- Added validation for user input in `drawAll` using zod, demo website now displays error for invalid inputs
 
 ### 🐛 Bug fixes